feat(visibility): scaffold the visibility context api (#427)
First stacked increment of the ClickHouse event-store work (ADR-0015).
Establishes the visibility bounded context's published language with no
behavior change:

- server/visibility/api: the EventLog (durable work queue) and EventArchive
  (event lake) interfaces every other context will consume, store-neutral and
  documented with the invariants implementations must uphold.
- The canonical Event envelope moves to visibility/api (it is visibility's to
  own); detection/api keeps `type Event = visibilityapi.Event` so all callers
  and the rules.api re-export stay byte-identical. Pure type alias, wire shape
  unchanged (detection/api wire round-trip tests pass).
- arch-go: new **.visibility.api block; **.detection.api may import
  **.visibility.api for the alias (dependency points detection -> visibility).

No runtime wiring, no new dependency, no datastore yet: the ClickHouse archive,
the MySQL event_queue, migrations, and the intake/processor changes land in the
functional-switch PR. Verified: go build ./..., arch-go, golangci-lint, and the
detection/rules unit suites are green.
